add: bikin handler

+++ pertemuan-ke-14-assigment-2/microservice-user-wallet-backup/entity/user_saldo.go
@@ -5,7 +5,7 @@ import "time"
 type UserSaldo struct {
 	Id        int       `gorm:"primaryKey" json:"id"`
 	UserId    int       `gorm:"type:int;not null" json:"name" binding:"required"`
-	Saldo     float32   `gorm:"type:float"`
+	Saldo     float64   `gorm:"type:float"`
 	CreatedAt time.Time `json:"created_at"`
 	UpdatedAt time.Time `json:"updated_at"`
 }
