+++ pertemuan-ke-14-assigment-2/microservice-user-wallet-backup/entity/user_saldo_history.go
@@ -8,7 +8,7 @@ type UserSaldoHistory struct {
 	UserIdTo        int       `gorm:"type:int;not null" json:"user_id_to" binding:"required"`
 	TypeTransaction string    `gorm:"type:string;not null" json:"type_transcation" binding:"required"`
 	TypeCredit      string    `gorm:"type:string;not null" json:"type_credit" binding:"required"`
-	Total           float32   `gorm:"type:float" json:"total" binding:"required`
+	Total           float64   `gorm:"type:float" json:"total" binding:"required`
 	CreatedAt       time.Time `json:"created_at"`
 	UpdatedAt       time.Time `json:"updated_at"`
 }
